% Generated by roxygen2 (4.0.2): do not edit by hand
\name{pdf.music}
\alias{pdf.music}
\title{makes a pdf file for the given piece of music}
\usage{
pdf.music(csv.file.name, pdf.file.name = "", plot.title = "",
  displayisolates = F, displaylabels = T)
}
\description{
makes a pdf file for the given piece of music
}

