% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plot.music}
\alias{plot.music}
\title{plots the network of the specific track of the music piece}
\usage{
\method{plot}{music}(csv.file.name, trackno, plot.title = "",
  displayisolates = F, displaylabels = T)
}
\description{
plots the network of the specific track of the music piece
}

